* automation now works correctly
Parameters which must be automated as a group are identified as such and their values collected each time through the automation loop.
At the end of the loop the collected parameters are then applied to the group's function and then the collected values removed from the collection so the loop can begin again.